feat: rbac ;

+++ internal/pkg/service/rbac/role/role_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestListByIDs(t *testing.T) {
 	list, _ := Server.List()
-	ls := ListByIDs([]int{1, 2}, list)
+	ls := ListByIDs([]int64{1, 2}, list)
 	for _, v := range ls {
 		log.Printf("v=%+v\n", v)
 	}
@@ -15,6 +15,6 @@ func TestListByIDs(t *testing.T) {
 
 func TestListPowerIDsByIDs(t *testing.T) {
 	list, _ := Server.List()
-	ls := ListPowerIDsByIDs([]int{1, 2}, list)
+	ls := ListPowerIDsByIDs([]int64{1, 2}, list)
 	log.Printf("v=%+v\n", ls)
 }
